requirements.txt
discord.py>=2.3.0
aiohttp>=3.9.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
asyncio>=3.4.3
typing-extensions>=4.8.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
openai>=1.0.0
